# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  modelHelper.AddressInput:
    properties:
      city:
        type: string
      district:
        type: string
      house_number:
        type: string
      landmark:
        type: string
      pincode:
        type: string
      street:
        type: string
    type: object
...
  modelHelper.BlockUser:
    properties:
      reason:
        type: string
      user_id:
        type: integer
    type: object
...
  modelHelper.UserDataInput:
    properties:
      email:
        type: string
      f_name:
        type: string
      l_name:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - password
    - phone
    type: object
  modelHelper.UserLoginEmail:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  modelHelper.UserLoginPhone:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Ecommerce Web Application built using Go Lang, PSQL, REST API following
    Clean Architecture.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Ecommerce Web API
  version: "1.0"
paths:
  /addresses/:
    post:
      consumes:
      - application/json
      description: Add address
      operationId: add-address
      parameters:
      - description: User address
        in: body
        name: user_address
        required: true
        schema:
          $ref: '#/definitions/modelHelper.AddressInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can add address
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update address
      operationId: update-address
      parameters:
      - description: User address
        in: body
        name: user_address
        required: true
        schema:
          $ref: '#/definitions/modelHelper.AddressInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can update existing address
      tags:
      - Users
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Admin can list all registered users
      operationId: list-all-users
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: limit
        type: integer
      - description: Search query string
        in: query
        name: query
        type: string
      - description: Filter criteria for the users
        in: query
        name: filter
        type: string
      - description: Sorting criteria for the users
        in: query
        name: sort_by
        type: string
      - description: Sorting in descending order
        in: query
        name: sort_desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can list all registered users
      tags:
      - Admin
  /admin/users/{id}:
    get:
      consumes:
      - application/json
      description: Admin can fetch a specific user details using user id
      operationId: find-user-by-id
      parameters:
      - description: ID of the user to be fetched
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can fetch a specific user details using user id
      tags:
      - Admin
  /admin/users/block:
    put:
      consumes:
      - application/json
      description: Admin can block a registered user
      operationId: block-user
      parameters:
      - description: ID of the user to be blocked
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/modelHelper.BlockUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can block a user
      tags:
      - Admin
  /admin/users/unblock/{id}:
    put:
      consumes:
      - application/json
      description: Admin can unblock a blocked user
      operationId: unblock-user
      parameters:
      - description: ID of the user to be unblocked
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can unblock a blocked user
      tags:
      - Admin
...
  /login/email:
    post:
      consumes:
      - application/json
      description: Login as a user to access the ecommerce site
      operationId: user-login-email
      parameters:
      - description: User details
        in: body
        name: user_details
        required: true
        schema:
          $ref: '#/definitions/modelHelper.UserLoginEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login
      tags:
      - Users
  /login/phone:
    post:
      consumes:
      - application/json
      description: Login as a user to access the ecommerce site
      operationId: user-login-phone
      parameters:
      - description: User details
        in: body
        name: user_details
        required: true
        schema:
          $ref: '#/definitions/modelHelper.UserLoginPhone'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login-Phone
      tags:
      - Users
  /logout:
    get:
      consumes:
      - application/json
      description: Logs out a logged-in user from the E-commerce web api
      operationId: user-logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: User Logout
      tags:
      - Users
...
  /profile:
    get:
      consumes:
      - application/json
      description: Users can visit their profile
      operationId: user-profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can view their profile
      tags:
      - Users
...
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user with the specified details.
      operationId: create-user
      parameters:
      - description: User details
        in: body
        name: user_details
        required: true
        schema:
          $ref: '#/definitions/modelHelper.UserDataInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new user
      tags:
      - Users
...
swagger: "2.0"